Fix handling of loot chests * Relocate ChestManager (it is specialized to tiered chests) * Remove ChestListener; migrate fixed functionality to GameListener#onLootChestOpen * Update references (Commands, Survival) * Closes #50
