+++ src/main/java/dev/dpvb/survival/game/GameListener.java
@@ -1,6 +1,7 @@
 package dev.dpvb.survival.game;
 
 import dev.dpvb.survival.chests.airdrop.AirdropManager;
+import dev.dpvb.survival.chests.tiered.ChestManager;
 import dev.dpvb.survival.stats.PlayerInfoManager;
 import dev.dpvb.survival.util.messages.Messages;
 import org.bukkit.Material;
@@ -25,6 +26,25 @@ public class GameListener implements Listener {
 
     public GameListener(GameManager manager) {
         this.manager = manager;
+    }
+
+    // Handle loot chests (route player to virtual inventory)
+    @EventHandler
+    public void onLootChestOpen(PlayerInteractEvent event) {
+        if (event.getHand() != EquipmentSlot.HAND) return;
+        if (event.getAction() != Action.RIGHT_CLICK_BLOCK) return;
+        final var block = event.getClickedBlock();
+        //noinspection ConstantConditions (We know there is a block)
+        switch (block.getType()) {
+            case CHEST, ENDER_CHEST -> {
+                if (block.getWorld() != manager.getArenaWorld()) return;
+                final var lootChest = ChestManager.getInstance().getLootChestMap().get(block.getLocation());
+                if (lootChest != null) {
+                    event.setCancelled(true);
+                    lootChest.open(event.getPlayer());
+                }
+            }
+        }
     }
 
     @EventHandler
